services/parser_service/requirements.txt
# Parser Service Requirements
# Inherits from parent requirements.txt, adds FastAPI dependencies

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.12.0
python-multipart>=0.0.6
